src/background-object-selector/requirements.txt
pillow
numpy
scipy
rembg